# openapi.yaml
openapi: 3.0.3
info:
  title: FutMarket API
  description: REST API for FUTMARKET.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /players:
    get:
      summary: Player list
      description: Returns the player catalog with optional filters and pagination.
      parameters:
        - in: query
          name: league
          schema:
            $ref: '#/components/schemas/League'
          description: Filter by league.
        - in: query
          name: team
          schema:
            type: string
          description: Filter by team.
        - in: query
          name: playerPosition
          schema:
            $ref: '#/components/schemas/PlayerPosition'
          description: Filter by position.
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed).
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Number of items per page.
        - in: query
          name: sort
          schema:
            type: string
            default: name
          description: Field to sort by (e.g., name, currentTokenPrice).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePlayerDTO'
  /players/ranking:
    get:
      summary: Player ranking
      description: Returns the ranking based on current price with pagination.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed).
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Number of items per page.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePlayerDTO'
  /players/{id}:
    get:
      summary: Player details
      description: Returns full player details by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Player ID.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDTO'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /players/{id}/quotes:
    get:
      summary: Quote history
      description: Returns historical quotes for a player.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Player ID.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quote'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /players/stats/sync/{league}:
    post:
      summary: Sync stats by league
      description: Syncs player stats for a specific league.
      parameters:
        - in: path
          name: league
          required: true
          schema:
            $ref: '#/components/schemas/League'
      responses:
        '202':
          description: Accepted
...
  /quotes/player/{playerId}:
    get:
      summary: Quotes by player
      description: Returns quotes for a player.
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteDTO'
...
components:
  schemas:
    League:
      type: string
      enum: [PL, PD, FL1, BL1, SA]
    PlayerPosition:
      type: string
      enum: [FORWARD, MIDFIELDER, DEFENDER, GOALKEEPER]
...
    ValuationStrategy:
      type: string
      enum: [GENERAL_PERFORMANCE, POSITION_WEIGHTED]
    PagePlayerDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PlayerDTO'
          description: List of players for the current page.
        totalPages:
          type: integer
          description: Total number of pages.
        totalElements:
          type: integer
          description: Total number of elements.
        currentPage:
          type: integer
          description: Current page number (0-indexed).
        pageSize:
          type: integer
          description: Number of elements per page.
        hasNext:
          type: boolean
          description: Whether there is a next page.
        hasPrevious:
          type: boolean
          description: Whether there is a previous page.
      required: [content, totalPages, totalElements, currentPage, pageSize, hasNext, hasPrevious]
...
    PlayerDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        league:
          type: string
          description: League full name.
        team:
          type: string
        playerPosition:
          $ref: '#/components/schemas/PlayerPosition'
        currentTokenPrice:
          type: number
          format: double
        goals:
          type: number
          format: double
        assists:
          type: number
          format: double
        shots:
          type: number
          format: double
        keyPasses:
          type: number
          format: double
        dribbles:
          type: number
          format: double
        tackles:
          type: number
          format: double
        interceptions:
          type: number
          format: double
        rating:
          type: number
          format: double
        lastStatsSync:
          type: string
          format: date-time
          nullable: true
        externalId:
          type: string
      required: [id, name, league, team, playerPosition, currentTokenPrice]
    QuoteDTO:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        currentTokenPrice:
          type: number
          format: double
        strategy:
          $ref: '#/components/schemas/ValuationStrategy'
        score:
          type: number
          format: double
        calculatedAt:
          type: string
          format: date-time
      required: [id, playerId, currentTokenPrice, strategy, score, calculatedAt]
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        team:
          type: string
        league:
          type: string
        playerPosition:
          $ref: '#/components/schemas/PlayerPosition'
        issuedTokens:
          type: integer
        currentTokenPrice:
          type: number
          format: double
        goals:
          type: number
          format: double
        assists:
          type: number
          format: double
        shots:
          type: number
          format: double
        keyPasses:
          type: number
          format: double
        dribbles:
          type: number
          format: double
        tackles:
          type: number
          format: double
        interceptions:
          type: number
          format: double
        rating:
          type: number
          format: double
        externalId:
          type: integer
      required: [id, name, team, league, playerPosition, availableTokens, currentTokenPrice]
    Quote:
      type: object
      properties:
        id:
          type: integer
        player:
          $ref: '#/components/schemas/Player'
        currentTokenPrice:
          type: number
          format: double
        strategy:
          $ref: '#/components/schemas/ValuationStrategy'
        score:
          type: number
          format: double
        calculatedAt:
          type: string
          format: date-time
      required: [id, player, currentTokenPrice, strategy, score, calculatedAt]
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
          nullable: true
      required: [timestamp, status, error, message]
...